requirements.txt
anytree==2.13.0
asttokens==3.0.1
build123d==0.10.0
cadquery-ocp==7.8.1.1.post1
cadquery-ocp-proxy==7.9.3.1
cadquery_vtk==9.3.1
colorama==0.4.6
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
decorator==5.2.1
executing==2.2.1
ezdxf==1.4.3
fonttools==4.62.1
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kiwisolver==1.5.0
lib3mf==2.5.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
numpy==2.4.3
numpy-stl==3.2.0
ocp_gordon==0.2.0
ocpsvg==0.5.0
packaging==26.0
parso==0.8.6
pillow==12.1.1
prompt_toolkit==3.0.52
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-utils==3.9.1
scipy==1.17.1
six==1.17.0
stack-data==0.6.3
svgelements==1.9.6
svgpathtools==1.7.2
svgwrite==1.4.3
sympy==1.14.0
traitlets==5.14.3
trianglesolver==1.2
typing_extensions==4.15.0
wcwidth==0.6.0
webcolors==24.8.0
